Add stack, resolve warnings, get test suite running
There were a few problems with the test suite:

1. It was maybe needlessly using smallcheck to simply run a bunch of
IO tests. At the moment this is incredibly slow, dubiously useful and
no manner of passing flags to the executable seems to let you scope
down the test size effectively.
2. There was a test hanging forever because it used cancel instead of
cancelLinked, when unblocked, sometimes this test *fails*. I'm pretty
sure there's a fundamental bug here.

Our renewable locking just does regular locking with timeouts (writing
the expiry time into the key), but adds a locking layer on top which
locks access to the key itself. We also make heavy use of a blocking
hold on that outer key. If you have a process rightfully holding a key
and a process trying to get in, the process trying to get in by
holding the outer lock open for a fixed 5 seconds, if the lock expiry
is smaller than that, the renewing process will be locked out
from *renewing* and the competing thread will succeed, stealing the
lock.

You could proably fix the *test* by making the lock time itself > 5s
and make it a documented invariant that this thing just doesn't work
for lock durations <= 5s, but that isn't super satisfactory.