Use simpler native lock
Since Redis 2.6.12, set operations have all that we need for a
suitable, non-haskell-specific atomic locking mechanism. This is not
as robust as Redlock (https://redis.io/topics/distlock) which is
designed for multi-master environments, but our locking mechanism
before didn't support that either. This deprecates the renewable
variants of the locking mechanisms as they are simply aliases to the
simpler version. This new mechanism also avoids some edge cases that
were present before and should generally be more performant.