Don't let blocklock hold connections
Unfortunately, hedis does not expose enough internals to let us
control the queue, so we had to do a type change so the connection
is passed in and pulled from the pool each time it checks, thus
preventing it from monopolizing the connection for the entire
duration. Without this, it was very easy for users to implement an
application-level lock, specify a long wait and not realize they'd be
pulling the connection out of the pool for that whole time. With
default settings, that's 50 connections, so if this happens in a
highly concurrent application more than 50 times on a long-held lock,
things start to hang.