Define semantics of locking with 0 or negative timeout
It was unclear what behavior should be expected from a lock with a 0
or negative timeout, so the library was basically leaning on redis'
behavior, which is to error and not take the lock. I think it is a bit
more sensible to take the lock even if it will immediately expire.